fashion_mnist/requirements.txt
torchvision==0.15.2
tensorboard==2.14.0
six==1.16.0
